refactor: decompose cloud form modal into modular components
